Add new radio button and vacancy_options helper
