+++ config/locales/en.yml
@@ -150,6 +150,7 @@ en:
         job_location_options:
           at_one_school: At one school in the trust
           central_office: At the trust's head office
+          at_multiple_schools: At more than one school in the trust
         job_location_html: Where will the job be based? (<span class='text-red'>Required</span>)
       school_form:
         job_location_html: In which school will the job be based? (<span class='text-red'>Required</span>)
